+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 # Changelog
 
-## [Unreleased]
+## [6.2.1] 2024-02-27
 
 ### Added
 
-- [TD-6436] Change timeout to use the environment variable
+- [TD-6436] All Dataloaders use `DB_TIMEOUT_MILLIS` environment variable for timeout
 
 ## [6.2.0] 2024-02-26